+++ src/Components/Charts/TokenChart.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="TokenChart" ref="container" >
+
     <div class="legend">
       <div 
         class="legend-item" 
@@ -18,58 +19,90 @@
 
     <div class="time-range">
       Showing data from
-      <span class="date">{{ fromDate }}</span>
+
+      <v-menu>
+        <template v-slot:activator="{ on }">
+          <v-btn 
+            depressed
+            class="ds-button-tall ma-0 mb-2 fromDate"
+            v-on="on"
+          >
+            {{ fromDate }}
+          </v-btn>
+        </template>
+
+        <v-date-picker 
+          no-title
+          @change="setStartDate"
+          :allowedDates="isAllowedStartDate"
+        ></v-date-picker>
+      </v-menu>
+
       to
-      <span class="date">{{ toDate }}</span>
+
+      <v-menu>
+        <template v-slot:activator="{ on }">
+          <v-btn 
+            depressed
+            class="ds-button-tall ma-0 mb-2 toDate"
+            v-on="on"
+          >
+            {{ toDate }}
+          </v-btn>
+        </template>
+
+        <v-date-picker 
+          no-title
+          @change="setEndDate"
+          :allowedDates="isAllowedEndDate"
+        ></v-date-picker>
+      </v-menu>
     </div>
 
-    <svg :id="plotId">
-      <defs>
-        <clipPath id="clip">
-          <rect
-            width="500"
-            height="300"
-          />
-        </clipPath>
-      </defs>
-
-      <g class="y-axis" />
-      <g class="x-axis" />
-      <g
-        class="chart"
-        clip-path="url(#clip)"
-      />
-
-      <g class="context-y-axis" />
-      <g class="context-x-axis" />
-      <g
-        class="context-chart"
-        clip-path="url(#clip)"
-      />
-
-      <g
+    <div class="chart-container">
+      <div 
         class="tooltip"
         style="display: none"
       >
-        <rect
-          height="25"
-          fill="white"
-          stroke="#BBB"
-          stroke-width="0.5px"
+        <div
+          v-for="feature in features"
+          :class="feature.slug" 
+          :key="feature.slug"
+        > 
+        </div>
+
+        <div class="cumulative"></div>
+      </div>
+      <svg :id="plotId">
+        <defs>
+          <clipPath id="clip">
+            <rect
+              width="500"
+              height="300"
+            />
+          </clipPath>
+        </defs>
+
+        <g class="y-axis" />
+        <g class="x-axis" />
+        <g
+          class="chart"
+          clip-path="url(#clip)"
         />
-        <text
-          dx="8px"
-          dy="1.5em"
-          font-size="12px"
-          font-weight="bold"
+
+        <g class="context-y-axis" />
+        <g class="context-x-axis" />
+        <g
+          class="context-chart"
+          clip-path="url(#clip)"
         />
-      </g>
-    </svg>
+      </svg>
+    </div>
   </div>
 </template>
 
 
-<style>
+<style lang="scss">
 .TokenChart {
   display: inline-block;
   position: relative;
@@ -121,10 +154,14 @@
   overflow: hidden;
 }
 
-.TokenChart > svg {
+.TokenChart > .chart-container {
+  position: relative;
+}
+
+.TokenChart > .chart-container > svg {
   display: block;
   overflow: visible;
-  height: 500px;
+  height: 650px;
 }
 
 .TokenChart > svg .selection {
@@ -169,6 +206,34 @@
 
 .TokenChart .tooltip {
   z-index: 9999;
+  display: flex;
+  flex-direction: column-reverse;
+  width: fit-content;
+  position: absolute;
+  background: white;
+  border: 2px solid #e6e6e6;
+  box-shadow: 2px 5px 15px 0px rgba(black, 0.1);
+
+  * {
+    width: fit-content;
+    display: flex;
+    align-items: center;
+    font-size: 12px;
+    font-weight: 600;
+    color: #333;
+    padding: 8px 16px;
+    width: 100%;
+    box-sizing: border-box;
+
+    &:not(:first-of-type) {
+      border-bottom: 1px solid #eee;
+    }
+  }
+}
+
+.TokenChart .toDate,
+.TokenChart .fromDate {
+  margin: 0 0.5rem !important;
 }
 </style>
 
@@ -176,6 +241,8 @@
 <script>
 import * as d3 from 'd3';
 import * as moment from 'moment';
+import slugify from '../../core/slugify';
+import { DaySpan, Day } from 'dayspan';
 
 
 const DAY_IN_MS = 24 * 60 * 60 * 1000;
@@ -220,30 +287,38 @@ export default {
       min: 0,
       max: 0,
     },
-    cummulativeExtent: {
-      min: 0,
-      max: 0,
-    },
     focusMargin: {
       top: 40,
       right: 50,
-      bottom: 130,
+      bottom: 180,
       left: 30,
     },
 
     contextMargin: {
-      top: 330,
-      bottom: 0,
+      top: 500,
+      bottom: 30,
     },
   }),
-
   computed: {
-    fromDate() {
-      return moment.utc(this.focusExtent[0]).format('ddd, D MMM YYYY');
+    fromDate: {
+      cache: false,
+      get() {
+        return moment.utc(this.focusExtent[0]).format('ddd, D MMM YYYY');
+      },
     },
-    toDate() {
-      return moment.utc(this.focusExtent[1]).format('ddd, D MMM YYYY');
+    toDate: {
+      cache: false,
+      get() {
+        return moment.utc(this.focusExtent[1]).format('ddd, D MMM YYYY');
+      },
     },
+    today() {
+      return moment()
+    },
+  },
+
+  created() {
+    this.features.forEach(feat => feat.slug = slugify(feat.name.en));
   },
 
   /**
@@ -269,6 +344,11 @@ export default {
    * Component methods.
    */
   methods: {
+    /**
+     * Update the token data to be formatted correctly.
+     *
+     * @returns {void}
+     */
     formatTokenData() {
       const { data } = this;
 
@@ -285,6 +365,57 @@ export default {
         }
       }
     },
+    /**
+     * Updates the start date for the focused time range.
+     *
+     * @param {string} date the new start date.
+     * @returns {void}
+     */
+    setStartDate(date) {
+      this.focusExtent[0] = moment.utc(date).toDate();
+      this.render();
+    },
+
+    /**
+     * Updates the end date for the focused time range.
+     *
+     * @param {string} date the new end date.
+     * @returns {void}
+     */
+    setEndDate(date) {
+      this.focusExtent[1] = moment
+        .utc(date)
+        .add(15, 'hours')
+        .toDate();
+      this.render();
+    },
+
+    /**
+     * Checks whether the given date should be enabled.
+     *
+     * @param {string} date a given date.
+     * @return {boolean} whether this options should be enabled.
+     */
+    isAllowedStartDate(date) {
+      return (
+        (moment.utc(date) < this.focusExtent[1]) && 
+        (moment.utc(date) > ONE_MONTH_AGO)
+      );
+    },
+
+    /**
+     * Checks whether the given date should be enabled.
+     *
+     * @param {string} date a given date.
+     * @return {boolean} whether this options should be enabled.
+     */
+    isAllowedEndDate(date) {
+      return (
+        (moment.utc(date) > this.focusExtent[0]) && 
+        (moment.utc(date) <= moment.utc())
+      );
+    },
+
     contextBarWidth() {
       return this.width / 30 * 2/3;
     },
@@ -351,29 +482,26 @@ export default {
 
     computeValueExtent() {
       let cummulative;
-      let positive;
-      let negative;
-      let dataPoint;
-      let feature;
       let value;
 
-      for (let i = 0; i < this.formattedData.length; i++) {
-        positive = 0; 
-        negative = 0;
-        dataPoint = this.formattedData[i];
+      // Compute the maximum value of one user response.
+      for (let i = 0; i < this.features.length; i++) {
+        value = this.features[i].value;
 
-        for (let j = 0; j < this.features.length; j++) {
-          feature = this.features[j].name.en;
-          value = dataPoint[feature];
-
-          if (value === undefined) continue;
-
-          if (value < 0) {
-            negative += value;
-          } else {
-            positive += value;
-          }
+        if (value < 0) {
+          this.divergingExtent.min += value;
+        } else {
+          this.divergingExtent.max += value;
         }
+      }
+
+      let positive;
+      let negative;
+
+      // Find the maximum value for cumulative user responses.
+      for (let i = 0; i < this.data.length; i++) {
+        positive = this.data[i].positive;
+        negative = this.data[i].negative;
 
         if (positive > this.divergingExtent.max) {
           this.divergingExtent.max = positive;
@@ -381,14 +509,7 @@ export default {
 
         if (negative < this.divergingExtent.min) {
           this.divergingExtent.min = negative;
-        }
-
-        if (dataPoint.cummulative < this.cummulativeExtent.min) {
-          this.cummulativeExtent.min = dataPoint.cummulative;
-        }
-        if (dataPoint.cummulative > this.cummulativeExtent.max) {
-          this.cummulativeExtent.max = dataPoint.cummulative;
-        }
+        } 
       }
 
       if (this.divergingExtent.max % 2) {
@@ -403,8 +524,8 @@ export default {
     resize() {
       const dimensions = this.$refs.container.getBoundingClientRect();
       this.width = dimensions.width - this.focusMargin.left - this.focusMargin.right;
-      this.focusHeight = 450 - this.focusMargin.top - this.focusMargin.bottom;
-      this.contextHeight = 450 - this.contextMargin.top - this.contextMargin.bottom;
+      this.focusHeight = 650 - this.focusMargin.top - this.focusMargin.bottom;
+      this.contextHeight = 650 - this.contextMargin.top - this.contextMargin.bottom;
 
       // Set dimensions.
       this.svg
@@ -478,8 +599,8 @@ export default {
       const xAxis = d3
         .axisBottom()
         .scale(this.x)
-        .ticks(d3.utcDay)
         .tickSize(this.focusHeight)  // Height of the tick line.
+        .ticks(d3.timeDay)
         .tickFormat(d => moment.utc(d).format('MMM D'));
       const contextXAxis = d3
         .axisBottom()
@@ -487,17 +608,20 @@ export default {
         .ticks(30)
         .tickSize(this.contextHeight)  // Height of the tick line.
         .tickFormat(d => moment.utc(d).format('M/D'));
+      const yAxisTicks = this.y.ticks().filter(Number.isInteger);
       const yAxis = d3
         .axisLeft()
         .scale(this.y)
-        .ticks(Math.abs(this.divergingExtent.min) + this.divergingExtent.max + 1)
+        .tickValues(yAxisTicks)
         .tickSize(-this.width - focusBarWidth)  // Width of the tick line.
+        .ticks(this.divergingExtent.max)
         .tickFormat(d3.format('d'));
+      const contextYAxisTicks = this.contextY.ticks().filter(Number.isInteger);
       const contextYAxis = d3
         .axisLeft()
+        .tickValues(contextYAxisTicks)
         .scale(this.contextY)
         .tickSize(-this.width - focusBarWidth) // Width of the tick line.
-        .ticks(Math.abs(this.divergingExtent.min) + this.divergingExtent.max + 1)
         .tickFormat(d3.format('d'));
 
       // Append the axes.
@@ -518,9 +642,9 @@ export default {
         .select('.x-axis')
         .append('line')
         .style('stroke', '#efefef')
-        .style('stroke-width', 3)
+        .style('stroke-width', 2)
         .attr('class', 'base-line')
-        .attr('x1', 0)
+        .attr('x1', -focusBarWidth/2)
         .attr('x2', this.width + focusBarWidth/2)
         .attr('y1', this.y(0))
         .attr('y2', this.y(0));
@@ -555,13 +679,13 @@ export default {
      * @return {void}
      */
     drawFocusChart() {
-      const { svg, x, y, formattedData, features } = this;
+      const { svg, x, y, formattedData, focusMargin, features } = this;
       const barWidth = this.focusBarWidth();
       const stack = d3.stack()
         .keys(features.map(f => f.name.en))
         .offset(d3.stackOffsetDiverging);
       const layers = stack(formattedData);
-      const tooltip = svg.select('.tooltip');
+      const tooltip = document.querySelector('.TokenChart .tooltip');
 
       svg
         .select('.chart')
@@ -591,7 +715,7 @@ export default {
         .join('rect')
 
         // Set the bar position and dimension.
-        .attr('x', d => x(d.data.date))
+        .attr('x', d => x(d.data.date) + barWidth/2)
         .attr('width', s => barWidth)
         .attr('y', d => y(d[1]))
         .attr('height', d => {
@@ -599,24 +723,40 @@ export default {
         })
 
         // Tooltip
-        .on('mouseover', () => tooltip.style('display', null))
-        .on('mouseout', () => tooltip.style('display', 'none'))
+        .on('mouseover', () => tooltip.style.display = 'flex')
+        .on('mouseout', () => tooltip.style.display = 'none')
         .on('mousemove', function(d) {
           const el = d3.select(this);
-          const x = d3.mouse(d3.event.currentTarget)[0];
-          const y = d3.mouse(d3.event.currentTarget)[1] - 30;
+          const xCoords = x(d.data.date) + barWidth * 1.8;
+          const yCoords = y(d.data.positive);
+          const padding = 8;
+          const cumulativeLabel = 27;
           const date = d.data.date.toLocaleDateString(
             'default',
             { day: 'numeric', month: 'short' },
           );
 
-          tooltip.attr('transform', `translate(${x}, ${y})`);
-          const text = tooltip.select('text');
-          text.text(`${d.key}: ${d[1] - d[0]} `);
+          tooltip.style.left = xCoords + 'px'; 
+          tooltip.style.top = (yCoords - padding - cumulativeLabel) + 'px';
 
-          tooltip
-            .select('rect')
-            .attr('width', text.node().getBoundingClientRect().width + 16);
+          for (let i = 0; i < features.length; i++) {
+            const text = tooltip.querySelector(`.${features[i].slug}`)
+            const name = features[i].name.en;
+            const value = d.data[name];
+
+            if (!value) {
+              text.style.display = 'none';
+              continue;
+            }
+            
+            text.innerText = `${name}: ${value}`;
+            text.style.top = y(0) - y(value) + 'px';
+            text.style.height = y(0) - y(value) + 'px';
+            text.style.display = 'flex';
+          }
+
+          const text = tooltip.querySelector(`.cumulative`)
+          text.innerText = `Cumulative: ${d.data.cummulative}`;
         });
     },
 
@@ -628,6 +768,15 @@ export default {
         .select('.context-chart')
         .selectAll('.bar')
         .remove()
+      svg
+        .select('.context-chart')
+        .selectAll('.positive-bar')
+        .remove()
+      svg
+        .select('.context-chart')
+        .selectAll('.negative-bar')
+        .remove()
+
 
       // Negative
       svg
@@ -639,7 +788,7 @@ export default {
         .attr('class', 'negative-bar')
         .attr('fill', '#ED8495')
         // Set the bar position and dimension.
-        .attr('x', d => contextX(d.date))
+        .attr('x', d => contextX(d.date) + barWidth/2)
         .attr('width', barWidth)
         .attr('y', contextY(0))
         .attr('height', d => Math.abs(contextY(d.negative) - contextY(0)))
@@ -654,7 +803,7 @@ export default {
         .attr('class', 'positive-bar')
         .attr('fill', '#BEE0AC')
         // Set the bar position and dimension.
-        .attr('x', d => contextX(d.date))
+        .attr('x', d => contextX(d.date) + barWidth/2)
         .attr('width', barWidth)
         .attr('y', d => contextY(d.positive))
         .attr('height', d => Math.abs(contextY(d.positive) - contextY(0)))
@@ -670,7 +819,7 @@ export default {
         .attr('class', 'bar')
         .attr('fill', 'black')
         // Set the bar position and dimension.
-        .attr('x', d => contextX(d.date))
+        .attr('x', d => contextX(d.date) + barWidth/2)
         .attr('width', s => barWidth)
         .attr('y', d => contextY(Math.max(0, d.cummulative)))
         .attr('height', d => Math.abs(contextY(d.cummulative) - contextY(0)))