+++ src/models/Applet.js
@@ -74,7 +74,7 @@ export default class Applet {
       params: { users: JSON.stringify(users) },
     });
 
-    if( this.encryption ) {
+    if (this.encryption) {
       Applet.replaceItemValues(Applet.decryptResponses(data, this.encryption));
     }
 
@@ -116,8 +116,14 @@ export default class Applet {
       const responses = data.responses[itemId];
 
       for (let response of responses) {
-        if (response.value && response.value.ptr !== undefined && response.value.src !== undefined) {
-          response.value = data.dataSources[response.value.src].data[response.value.ptr];
+        if (
+          Array.isArray(response.value) && 
+          response.value.length > 0  && 
+          response.value[0].ptr !== undefined && 
+          response.value[0].src !== undefined
+        ) 
+        {
+          response.value = data.dataSources[response.value[0].src].data[response.value[0].ptr];
         }
       }
     }